Epic 31 P3 (4/4): delete the superseded GitHub-only surfaces; ratchet 18 -> 6
DELETED (zero remaining consumers after milestones 1-3):
- Tamma.Api: GitHubIntegrationService (1074-line live client absorbed into the
  driver in P1; delegator now dead), CIIntegrationService + CiClientFactory,
  IntegrationService (composite facade), the IGitHubEngineCallbackService seam
  (interface + Octokit + Null impls, incl. IRepoInstallationResolver), and
  OctokitGitHubActionsClient. All dead registrations dropped from Program.cs.
- Tamma.Activities: the IGitHubActionsClient seam (interface + Null impl).
  WorkflowRunSummary (engine-plane wire data AgentMonitorService loops over)
  moved to its own file.
- Tamma.Core: IGitHubIntegrationService / ICIIntegrationService / the composite
  IIntegrationService interfaces removed from IIntegrationService.cs; the wire
  DTO classes (GitHub*/TestRunResult/BuildStatus/Jira*) REMAIN - they are the
  mediation planes' response shapes and the Slack/JIRA/email interfaces stay.
- Result envelope kept: GitHubCallbackResult<T> + the engine result records
  moved to Services/Engine/EngineCallbackResults.cs (the deployed activities
  branch on the exact legacy 503 github_client_not_configured wire code).

KEPT for P4 (webhook/secrets seams 8-11, checked before deleting):
OctokitGitHubAppClient + IGitHubAppClient + NullGitHubAppClient +
GitHubInstallationServiceCollectionExtensions + InstallationRouterService +
LibsodiumGitHubSecretsProvisioner - install-time TAMMA_API_KEY provisioning
and libsodium CI-secrets still mint GitHub App installation tokens through
them; the Octokit package reference stays for exactly these files.

Ratchet: baseline 18 -> 6 (PinHistory [26, 21, 18, 6]) - the six remaining
entries are exactly the P4 webhook/secrets seam files. Cutover pins
(IntegrationServiceCutoverTests, AgentDispatchCutoverTests, thin-client tests)
converted from typeof() to NAME checks so they stay armed against
reintroduction of the deleted types.

Gates: build 0 errors; Activities.Tests FULL 3588 pass; Platforms suites
116/96/97/109/72 pass; Core.Tests 1035; Guardrails.Tests 32; Api.Tests broad
filter (Git|Ci|Engine|Platforms|AgentDispatch|Webhook|GitHub|governance
sweeps) 536 pass; ratchet green at 6.
